+++ api/src/main/java/io/strimzi/api/kafka/model/CruiseControlSpec.java
@@ -7,6 +7,7 @@ package io.strimzi.api.kafka.model;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 import io.fabric8.kubernetes.api.model.ResourceRequirements;
+import io.strimzi.api.annotations.DeprecatedProperty;
 import io.strimzi.api.kafka.model.balancing.BrokerCapacity;
 import io.strimzi.api.kafka.model.template.CruiseControlTemplate;
 import io.strimzi.crdgenerator.annotations.Description;
@@ -28,7 +29,7 @@ import lombok.EqualsAndHashCode;
         "livenessProbe", "readinessProbe",
         "jvmOptions", "logging",
         "template", "brokerCapacity",
-        "config", "metrics"})
+        "config", "metrics", "jmxExporterMetrics"})
 @EqualsAndHashCode
 public class CruiseControlSpec implements UnknownPropertyPreserving, Serializable {
     private static final long serialVersionUID = 1L;
@@ -52,6 +53,7 @@ public class CruiseControlSpec implements UnknownPropertyPreserving, Serializabl
     private BrokerCapacity brokerCapacity;
     private Map<String, Object> config = new HashMap<>(0);
     private Map<String, Object> metrics;
+    private JmxExporterMetrics jmxExporterMetrics;
     private Map<String, Object> additionalProperties = new HashMap<>(0);
 
     @Description("The docker image for the pods.")
@@ -96,6 +98,8 @@ public class CruiseControlSpec implements UnknownPropertyPreserving, Serializabl
         this.config = config;
     }
 
+    @DeprecatedProperty(movedToPath = "spec.cruiseControl.jmxExporterMetrics")
+    @Deprecated
     @JsonInclude(JsonInclude.Include.NON_EMPTY)
     @Description("The Prometheus JMX Exporter configuration. " +
             "See https://github.com/prometheus/jmx_exporter for details of the structure of this configuration.")
@@ -105,6 +109,17 @@ public class CruiseControlSpec implements UnknownPropertyPreserving, Serializabl
 
     public void setMetrics(Map<String, Object> metrics) {
         this.metrics = metrics;
+    }
+
+    @JsonInclude(JsonInclude.Include.NON_EMPTY)
+    @Description("The Prometheus JMX Exporter configuration. " +
+            "See https://github.com/prometheus/jmx_exporter for details of the structure of this configuration.")
+    public JmxExporterMetrics getJmxExporterMetrics() {
+        return jmxExporterMetrics;
+    }
+
+    public void setJmxExporterMetrics(JmxExporterMetrics jmxExporterMetrics) {
+        this.jmxExporterMetrics = jmxExporterMetrics;
     }
 
     @Description("Logging configuration (log4j1) for Cruise Control.")