+++ cluster-operator/src/main/java/io/strimzi/operator/cluster/operator/assembly/KafkaConnectS2IAssemblyOperator.java
@@ -115,8 +115,8 @@ public class KafkaConnectS2IAssemblyOperator extends AbstractConnectOperator<Ope
         ConfigMap loggingCm = connect.getLogging() instanceof ExternalLogging ?
                 configMapOperations.get(namespace, ((ExternalLogging) connect.getLogging()).getName()) :
                 null;
-        ConfigMap metricsCm = connect.isExternalMetricsConfigured() ?
-                configMapOperations.get(namespace, connect.getExternalMetricsName()) :
+        ConfigMap metricsCm = connect.isJmxExporterMetricsConfigured() ?
+                configMapOperations.get(namespace, connect.getJmxExporterMetrics().getValueFrom().getConfigMapKeyRef().getName()) :
                 null;
 
         ConfigMap logAndMetricsConfigMap = connect.generateMetricsAndLogConfigMap(loggingCm, metricsCm);