+++ CHANGELOG.md
@@ -10,7 +10,7 @@
 * Allow disabling ownerReference on CA secrets
 * Move from Docker Hub to Quay.io as our container registry
 * Add possibility to configure DeploymentStrategy for Kafka Connect, Kafka Mirror Maker (1 and 2), and Kafka Bridge
-* Support configuring metrics by a ConfigMap
+* Support passing metrics configuration as an external ConfigMap
 
 ## 0.20.0